tarea3.10 internacionalizacion español e inglés ;)

+++ src/main/java/org/iesvdm/dao/ComercialDAOImpl.java
@@ -44,7 +44,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 			ps.setString(idx++, comercial.getNombre());
 			ps.setString(idx++, comercial.getApellido1());
 			ps.setString(idx++, comercial.getApellido2());
-			ps.setDouble(idx++, comercial.getComision());
+			ps.setBigDecimal(idx++, comercial.getComision());
 			return ps;
 		},keyHolder);
 
@@ -63,7 +63,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 						rs.getString("nombre"),
 						rs.getString("apellido1"),
 						rs.getString("apellido2"),
-						rs.getFloat("comisión"))
+						rs.getBigDecimal("comisión"))
 
 		);
 
@@ -84,7 +84,7 @@ public class ComercialDAOImpl implements ComercialDAO {
 								rs.getString("nombre"),
 								rs.getString("apellido1"),
 								rs.getString("apellido2"),
-								rs.getDouble("comisión"))
+								rs.getBigDecimal("comisión"))
 						, id
 				);
