+++ src/main/java/org/iesvdm/dto/ComercialDTO.java
@@ -4,6 +4,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,6 +15,6 @@ public class ComercialDTO {
     private String nombre;
     private String apellido1;
     private String apellido2;
-    private double comisión;
+    private BigDecimal comisión;
 
 }
